requirements.txt
python-pkcs11==0.7.0
cryptography==3.1
